chore(example): move client example app to examples dir
