+++ .run/NostrClientExampleApplication.run.xml
@@ -1,0 +1,17 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="NostrClientExampleApplication" type="Application" factoryName="Application" nameIsGenerated="true">
+    <option name="MAIN_CLASS_NAME" value="org.tbk.nostr.example.client.NostrClientExampleApplication" />
+    <module name="nostr-spring-boot-starter.examples.nostr-client-example-application.main" />
+    <option name="PROGRAM_PARAMETERS" value="--spring.profiles.active=development" />
+    <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/examples/nostr-client-example-application" />
+    <extension name="coverage">
+      <pattern>
+        <option name="PATTERN" value="org.tbk.nostr.example.client.*" />
+        <option name="ENABLED" value="true" />
+      </pattern>
+    </extension>
+    <method v="2">
+      <option name="Make" enabled="true" />
+    </method>
+  </configuration>
+</component>
