--- .run/NostrExampleApplication.run.xml
@@ -1,17 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="NostrExampleApplication" type="Application" factoryName="Application" nameIsGenerated="true">
-    <option name="MAIN_CLASS_NAME" value="org.tbk.nostr.example.NostrExampleApplication" />
-    <module name="nostr-spring-boot-starter.nostr.nostr-example-application.main" />
-    <option name="PROGRAM_PARAMETERS" value="--spring.profiles.active=development" />
-    <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/nostr/nostr-example-application" />
-    <extension name="coverage">
-      <pattern>
-        <option name="PATTERN" value="org.tbk.nostr.example.*" />
-        <option name="ENABLED" value="true" />
-      </pattern>
-    </extension>
-    <method v="2">
-      <option name="Make" enabled="true" />
-    </method>
-  </configuration>
-</component>
