+++ examples/nostr-client-example-application/build.gradle
@@ -1,6 +1,6 @@
 apply plugin: 'org.springframework.boot'
 
-description = 'nostr example application package'
+description = 'nostr client example application package'
 
 dependencies {
     implementation project(':nostr:nostr-core')
