+++ examples/nostr-client-example-application/readme.md
@@ -1,0 +1,6 @@
+nostr-client-example-application
+===
+
+```shell
+./gradlew -p examples/nostr-client-example-application bootRun
+```
