+++ examples/nostr-client-example-application/src/main/resources/banner.txt
