+++ examples/nostr-client-example-application/src/main/resources/logback-development.xml
