--- nostr/nostr-example-application/readme.md
@@ -1,6 +1,0 @@
-nostr-example-application
-===
-
-```shell
-./gradlew -p nostr/nostr-example-application bootRun
-```
